refactoring to use both api's v1 and v2 at the same time
