+++ xchange-bitfinex/src/main/java/org/knowm/xchange/bitfinex/dto/BitfinexException.java
@@ -1,4 +1,4 @@
-package org.knowm.xchange.bitfinex.common.dto;
+package org.knowm.xchange.bitfinex.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
