+++ xchange-bitfinex/src/main/java/org/knowm/xchange/bitfinex/service/BitfinexAccountService.java
@@ -1,4 +1,4 @@
-package org.knowm.xchange.bitfinex.v1.service;
+package org.knowm.xchange.bitfinex.service;
 
 import java.io.IOException;
 import java.math.BigDecimal;
@@ -6,9 +6,8 @@ import java.util.Date;
 import java.util.List;
 import java.util.Map;
 import org.knowm.xchange.Exchange;
-import org.knowm.xchange.bitfinex.common.BitfinexErrorAdapter;
-import org.knowm.xchange.bitfinex.common.dto.BitfinexException;
-import org.knowm.xchange.bitfinex.v1.BitfinexAdapters;
+import org.knowm.xchange.bitfinex.BitfinexErrorAdapter;
+import org.knowm.xchange.bitfinex.dto.BitfinexException;
 import org.knowm.xchange.bitfinex.v1.BitfinexUtils;
 import org.knowm.xchange.bitfinex.v1.dto.account.BitfinexDepositAddressResponse;
 import org.knowm.xchange.currency.Currency;
