+++ xchange-bitfinex/src/main/java/org/knowm/xchange/bitfinex/service/BitfinexMarketDataService.java
@@ -1,12 +1,14 @@
-package org.knowm.xchange.bitfinex.v1.service;
+package org.knowm.xchange.bitfinex.service;
 
 import java.io.IOException;
+import java.util.Arrays;
 import java.util.Date;
 import java.util.List;
+import java.util.stream.Collectors;
+
 import org.knowm.xchange.Exchange;
-import org.knowm.xchange.bitfinex.common.BitfinexErrorAdapter;
-import org.knowm.xchange.bitfinex.common.dto.BitfinexException;
-import org.knowm.xchange.bitfinex.v1.BitfinexAdapters;
+import org.knowm.xchange.bitfinex.BitfinexErrorAdapter;
+import org.knowm.xchange.bitfinex.dto.BitfinexException;
 import org.knowm.xchange.bitfinex.v1.BitfinexUtils;
 import org.knowm.xchange.bitfinex.v1.dto.marketdata.BitfinexDepth;
 import org.knowm.xchange.bitfinex.v1.dto.marketdata.BitfinexLendDepth;
@@ -20,6 +22,7 @@ import org.knowm.xchange.dto.trade.FixedRateLoanOrder;
 import org.knowm.xchange.dto.trade.FloatingRateLoanOrder;
 import org.knowm.xchange.exceptions.ExchangeException;
 import org.knowm.xchange.service.marketdata.MarketDataService;
+import org.knowm.xchange.service.marketdata.params.Params;
 
 /**
  * Implementation of the market data service for Bitfinex
@@ -43,9 +46,22 @@ public class BitfinexMarketDataService extends BitfinexMarketDataServiceRaw
 
   @Override
   public Ticker getTicker(CurrencyPair currencyPair, Object... args) throws IOException {
+    //return getTickerV1(currencyPair, args);
+    return getTickerV2(currencyPair, args);
+  }
+
+  private Ticker getTickerV1(CurrencyPair currencyPair, Object... args) throws IOException {
     try {
       return BitfinexAdapters.adaptTicker(
           getBitfinexTicker(BitfinexUtils.toPairString(currencyPair)), currencyPair);
+    } catch (BitfinexException e) {
+      throw BitfinexErrorAdapter.adapt(e);
+    }
+  }
+
+  private Ticker getTickerV2(CurrencyPair currencyPair, Object... args) throws IOException {
+    try {
+      return BitfinexAdapters.adaptTicker(getBitfinexTickerV2(currencyPair));
     } catch (BitfinexException e) {
       throw BitfinexErrorAdapter.adapt(e);
     }
@@ -128,14 +144,19 @@ public class BitfinexMarketDataService extends BitfinexMarketDataServiceRaw
     }
   }
 
+  @Override
+  public Trades getTrades(CurrencyPair currencyPair, Object... args) throws IOException {
+    //return getTradesV1(currencyPair, args);
+     return getTradesV2(currencyPair, args);
+  }
+
   /**
    * @param currencyPair The CurrencyPair for which to query trades.
    * @param args One argument may be supplied which is the timestamp after which trades should be
    *     collected. Trades before this time are not reported. The argument may be of type
    *     java.util.Date or Number (milliseconds since Jan 1, 1970)
    */
-  @Override
-  public Trades getTrades(CurrencyPair currencyPair, Object... args) throws IOException {
+  private Trades getTradesV1(CurrencyPair currencyPair, Object... args) throws IOException {
     try {
       long lastTradeTime = 0;
       if (args != null && args.length == 1) {
@@ -159,6 +180,63 @@ public class BitfinexMarketDataService extends BitfinexMarketDataServiceRaw
           getBitfinexTrades(BitfinexUtils.toPairString(currencyPair), lastTradeTime);
 
       return BitfinexAdapters.adaptTrades(trades, currencyPair);
+    } catch (BitfinexException e) {
+      throw BitfinexErrorAdapter.adapt(e);
+    }
+  }
+
+  /**
+   * @param currencyPair The CurrencyPair for which to query public trades.
+   * @param args Upto 4 numeric arguments may be supplied limitTrades, startTimestamp (Unix
+   *     millisecs), endTimestamp (Unix millisecs), sort -1 / 1 (if = 1 it sorts results returned
+   *     with old > new)
+   */
+  private Trades getTradesV2(CurrencyPair currencyPair, Object... args) throws IOException {
+    try {
+      int limitTrades = 1000;
+      long startTimestamp = 0;
+      long endTimestamp = 0;
+      int sort = -1;
+
+      if (args != null) {
+        for (int i = 0; i < args.length; i++) {
+          if (args[i] instanceof Number) {
+            Number arg = (Number) args[i];
+            switch (i) {
+              case 0:
+                limitTrades = arg.intValue();
+                break;
+              case 1:
+                startTimestamp = arg.longValue();
+                break;
+              case 2:
+                endTimestamp = arg.longValue();
+                break;
+              case 3:
+                sort = arg.intValue();
+                break;
+            }
+          } else {
+            throw new IllegalArgumentException(
+                "Extra argument #" + i + " must be an int/long was: " + args[i].getClass());
+          }
+        }
+      }
+
+      return BitfinexAdapters.adaptPublicTrades(
+          getBitfinexPublicTrades(currencyPair, limitTrades, startTimestamp, endTimestamp, sort),
+          currencyPair);
+    } catch (BitfinexException e) {
+      throw BitfinexErrorAdapter.adapt(e);
+    }
+  }
+
+  @Override
+  public List<Ticker> getTickers(Params params) throws IOException {
+    try {
+      return Arrays.stream(getBitfinexTickers())
+          .map(BitfinexAdapters::adaptTicker)
+          .collect(Collectors.toList());
     } catch (BitfinexException e) {
       throw BitfinexErrorAdapter.adapt(e);
     }
