+++ xchange-bitfinex/src/main/java/org/knowm/xchange/bitfinex/service/BitfinexPayloadDigest.java
@@ -1,4 +1,4 @@
-package org.knowm.xchange.bitfinex.common.service;
+package org.knowm.xchange.bitfinex.service;
 
 import java.util.Base64;
 import si.mazi.rescu.ParamsDigest;
