--- xchange-bitfinex/src/main/java/org/knowm/xchange/bitfinex/v2/service/BitfinexTradeService.java
@@ -1,17 +1,0 @@
-package org.knowm.xchange.bitfinex.v2.service;
-
-import org.knowm.xchange.Exchange;
-import org.knowm.xchange.bitfinex.v1.service.BitfinexTradeServiceRaw;
-import org.knowm.xchange.service.trade.TradeService;
-
-public class BitfinexTradeService extends BitfinexTradeServiceRaw implements TradeService {
-
-  /**
-   * Constructor
-   *
-   * @param exchange
-   */
-  public BitfinexTradeService(Exchange exchange) {
-    super(exchange);
-  }
-}
