+++ xchange-bitfinex/src/test/java/org/knowm/xchange/bitfinex/v2/BitfinexAdaptersTest.java
@@ -1,12 +1,11 @@
 package org.knowm.xchange.bitfinex.v2;
 
-import static org.junit.Assert.*;
-
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import org.junit.Assert;
 import org.junit.Test;
+import org.knowm.xchange.bitfinex.service.BitfinexAdapters;
 import org.knowm.xchange.currency.CurrencyPair;
 
 public class BitfinexAdaptersTest {
